backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
playwright
python-dotenv
google-cloud-storage
python-multipart
google-cloud-texttospeech